+++ g42cloud-sdk-core/g42cloudsdkcore/http/http_client.py
@@ -70,6 +70,10 @@ class HttpClient(object):
     def config(self):
         return self._config
 
+    @property
+    def logger(self):
+        return self._logger
+
     def do_request_sync(self, request):
         invoke = getattr(self._session, request.method.lower())
 
@@ -109,8 +113,9 @@ class HttpClient(object):
         fun = getattr(FutureSession(self._session, self._executor), request.method.lower())
         hooks.append(self.response_error_hook_factory())
 
+        url = "%s://%s%s" % (request.schema, request.host, request.uri)
         future = fun(
-            "%s://%s%s" % (request.schema, request.host, request.uri),
+            url,
             timeout=self._config.timeout,
             headers=request.header_params,
             proxies=self._proxy,
@@ -131,11 +136,7 @@ class HttpClient(object):
             try:
                 resp.raise_for_status()
             except HTTPError as httpError:
-                sdk_error = self._exception_handler.handle_exception(httpError.request, httpError.response)
-                if 400 <= httpError.response.status_code < 500:
-                    raise exceptions.ClientRequestException(httpError.response.status_code, sdk_error)
-                else:
-                    raise exceptions.ServerResponseException(httpError.response.status_code, sdk_error)
+                self._exception_handler.handle_exception(httpError.request, httpError.response)
             except Timeout as timeout:
                 raise exceptions.CallTimeoutException(str(timeout))
             except TooManyRedirects as tooManyRedirects: