+++ g42cloud-sdk-rds/g42cloudsdkrds/v3/rds_client.py
@@ -16,13 +16,16 @@ class RdsClient(Client):
 
     @classmethod
     def new_builder(cls, clazz=None):
-        if clazz is None:
-            return ClientBuilder(cls)
+        if not clazz:
+            client_builder = ClientBuilder(cls)
+        else:
+            if clazz.__name__ != "RdsClient":
+                raise TypeError("client type error, support client type is RdsClient")
+            client_builder = ClientBuilder(clazz)
 
-        if clazz.__name__ != "RdsClient":
-            raise TypeError("client type error, support client type is RdsClient")
+        
 
-        return ClientBuilder(clazz)
+        return client_builder
 
     def apply_configuration_async(self, request):
         """