+++ g42cloud-sdk-vpc/g42cloudsdkvpc/v3/vpc_client.py
@@ -16,13 +16,16 @@ class VpcClient(Client):
 
     @classmethod
     def new_builder(cls, clazz=None):
-        if clazz is None:
-            return ClientBuilder(cls)
+        if not clazz:
+            client_builder = ClientBuilder(cls)
+        else:
+            if clazz.__name__ != "VpcClient":
+                raise TypeError("client type error, support client type is VpcClient")
+            client_builder = ClientBuilder(clazz)
 
-        if clazz.__name__ != "VpcClient":
-            raise TypeError("client type error, support client type is VpcClient")
+        
 
-        return ClientBuilder(clazz)
+        return client_builder
 
     def batch_create_sub_network_interface(self, request):
         """